Revisit pathman readiness again.
Most pathman functions break if pathman is disabled: earlier I put defense
checks in split and merge range partitions, now it is
drop_range_partition_expand_next. Looks like the reason is that pathman caches
are not being invalidated if it is disabled: pathman_relcache_hook exits right
away then. This is kinda reasonable: if we want to disable pathman completely,
why maintain the caches? So this time try to bury the readiness check deeper,
in get_pathman_relation_info itself.

BTW, pathman caches are not dropped when it is disabled, which looks suspicious
on its own -- probably if we re-enable it later, caches might be inconsistent.