Unify unload_config and DisablePathman and do it (purge caches) on enable = false.
Probably there used to be an idea that with 'set pg_pathman = false' you can
just temporary disable planning hooks etc without resetting pathman caches,
but that never worked properly. At least, pathman_relcache_hook refuses to inval
cache if it is disabled. Also, if extension is disabled we never get to
unload_config, which means you could
 - disable pathman (guc disabled, but caches not destroyed)
 - drop extension (still caches are here)
 - create it back again, now cached relids are wrong

With some care we could totally separate them by maintaining caches even when
pathman is disabled, but is it worth it?