Add tableoid junk column to processed_tlist only for top-level parent.
(Note that the column is added to RelOptInfo reltarget of all nodes in the
hierarchy anyway; but only tle with top-level varno must get to *main* tlist,
i.e. processed_tlist.)

Previously it was added for each parent in the tree, i.e. multiple times in case
of multi-level partitioning, leading to

ERROR: variable not found in subplan target lists

errors in setref.

As comments say, this code better be rewritten to actually let parent deal with
its isParent flag. And to recurce with toplever rc. Probably should be done if
adjacent bugs arise.